Client-Server file exchange using UDP Sockets
Implemented Client-Server Networking system using UDP Sockets and base socket module in Python. Files and be stored and retrieved from the server by the client as and when required. 
Since it is UDP, to ensure reliability, an ACK system was also implemented to ensure safe transfer of data. Stop and Wait protocol was used for this purpose.